Tweak namespace autoloading
