+++ tests/bootstrap.php
@@ -3,11 +3,14 @@
 use VCR\VCR;
 use allejo\VCR\VCRCleaner;
 
-if (!file_exists('test/cassettes')) {
-    mkdir('test/cassettes', 0777, true);
+// require __DIR__ . '/../vendor/autoload.php';
+// require_once 'src/TuneupTechnology.php';
+
+if (!file_exists('tests/cassettes')) {
+    mkdir('tests/cassettes', 0777, true);
 }
 
-VCR::configure()->setCassettePath('test/cassettes')
+VCR::configure()->setCassettePath('tests/cassettes')
     ->setWhiteList(array('vendor/guzzle'))
     ->setStorage('yaml')
     ->setMode('once');
