+++ tests/cassettes/customers/delete.yml
