+++ tests/cassettes/inventory/all.yml
