+++ tests/cassettes/inventory/create.yml
