+++ tests/cassettes/inventory/delete.yml
