+++ tests/cassettes/inventory/retrieve.yml
