+++ tests/cassettes/inventory/update.yml
