+++ tests/cassettes/locations/all.yml
