+++ tests/cassettes/locations/delete.yml
