+++ tests/cassettes/locations/retrieve.yml
